chore: Create single constructor for LnDlcNode
Allows custom node creation.
Nodes can choose the policies they want to follow themselves (e.g. whether to
persist scorer data, or what LDK UserConfig to supply).

Note: app and coordinator configs are still in the same crate.
